components/markdown-processor/requirements.txt
PyPDF2
google-cloud-storage
google-genai